requirements.txt

Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.1
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0
django-storages==1.14.2
boto3==1.34.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.25.2
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
